Fix bug where too many scatters were generated
This happens when the estimated scatter size does not cleanly divide the
total size of the specified reference. In this case, you can end up with
an extra, very small scatter.

This commit fixes this issue by keeping track of the number of generated
scatters, and adding the regions of the additional scatter to the
previous one. This way, there will always be --scatter-count regions
generated.